Move TODO management to its own container component
